fixed sapling-output: previously, sapling output was limited to RUs that have saplings >1.3m, now output is written also if only small saplings are there.

+++ src/core/standstatistics.h
@@ -84,8 +84,8 @@ private:
     double mNPPabove;
     double mNPPsaplings; // carbon gain of saplings
     // regeneration layer
-    int mCohortCount; ///< number of cohrots
-    int mSaplingCount; ///< number of sapling (Reinekes Law)
+    double mCohortCount; ///< number of cohrots
+    double mSaplingCount; ///< number of sapling (Reinekes Law)
     double mSumSaplingAge;
     double mAverageSaplingAge;
     // carbon and nitrogen pools
